refactor: extract kftray-ssl crate
Move certificate generation, OS-native trust store integration, and the cert lifecycle manager out of kftray-portforward into a dedicated crate. Zero behaviour change. The new crate has no dependency on kube-portforward or Kubernetes.
