+++ crates/kftray-ssl/Cargo.toml
@@ -13,4 +13,31 @@ name = "kftray_ssl"
 path = "src/lib.rs"
 
 [dependencies]
-# Filled in 1.2 after moving code
+anyhow = { workspace = true }
+base64 = { workspace = true }
+dirs = { workspace = true }
+keyring-core = { workspace = true }
+kftray-commons = { workspace = true }
+lazy_static = { workspace = true }
+log = { workspace = true }
+pem = { workspace = true }
+rcgen = { workspace = true }
+rustls = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+time = { workspace = true }
+tokio = { workspace = true }
+
+[dev-dependencies]
+tempfile = { workspace = true }
+
+[target.'cfg(target_os = "linux")'.dependencies]
+dbus-secret-service-keyring-store = { workspace = true }
+linux-keyutils-keyring-store = { workspace = true }
+
+[target.'cfg(target_os = "macos")'.dependencies]
+apple-native-keyring-store = { workspace = true }
+
+[target.'cfg(target_os = "windows")'.dependencies]
+windows = { workspace = true }
+windows-native-keyring-store = { workspace = true }
