+++ crates/kftray-ssl/src/platform/macos.rs
