+++ crates/kftray-ssl/src/platform/windows.rs
