+++ crates/kftray-tauri/Cargo.toml
@@ -32,6 +32,7 @@ kftray-helper = { workspace = true }
 kftray-mcp = { workspace = true }
 kftray-network-monitor = { workspace = true }
 kftray-portforward = { workspace = true }
+kftray-ssl = { workspace = true }
 kftray-shortcuts = { workspace = true }
 kube = { workspace = true }
 lazy_static = { workspace = true }
