+++ crates/kftui/src/main.rs
@@ -29,8 +29,8 @@ use crate::logging::{
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
-    kftray_portforward::ssl::install_default_keyring_store();
-    kftray_portforward::ssl::ensure_crypto_provider_installed();
+    kftray_ssl::install_default_keyring_store();
+    kftray_ssl::ensure_crypto_provider_installed();
 
     let cli = Cli::parse();
